refactor: create own file for hoc
